+++ src/realm/object-store/sync/impl/sync_metadata.hpp
@@ -25,28 +25,11 @@
 
 #include <realm/obj.hpp>
 #include <realm/table.hpp>
-#include <realm/util/optional.hpp>
 #include <string>
 
 namespace realm {
+class SyncFileManager;
 class SyncMetadataManager;
-
-// A facade for a metadata Realm object representing app metadata
-class SyncAppMetadata {
-public:
-    struct Schema {
-        ColKey id_col;
-        ColKey deployment_model_col;
-        ColKey location_col;
-        ColKey hostname_col;
-        ColKey ws_hostname_col;
-    };
-
-    std::string deployment_model;
-    std::string location;
-    std::string hostname;
-    std::string ws_hostname;
-};
 
 // A facade for a metadata Realm object representing a sync user.
 class SyncUserMetadata {
@@ -207,6 +190,11 @@ class SyncMetadataManager {
     friend class SyncFileActionMetadata;
 
 public:
+    std::vector<SyncUserMetadata> all_logged_in_users() const;
+
+    // Perform all pending file actions and delete any remaining data for removed users.
+    void perform_launch_actions(SyncFileManager& file_manager) const;
+
     // Return a Results object containing all users not marked for removal.
     SyncUserMetadataResults all_unmarked_users() const;
 
@@ -225,6 +213,8 @@ public:
 
     // Retrieve file action metadata.
     util::Optional<SyncFileActionMetadata> get_file_action_metadata(StringData path) const;
+    // Perform any stored file actions for the given path.
+    bool perform_file_actions(SyncFileManager& file_manager, StringData path) const;
 
     // Create file action metadata.
     void make_file_action_metadata(StringData original_name, StringData partition_key_value, StringData local_uuid,
@@ -232,17 +222,6 @@ public:
 
     util::Optional<std::string> get_current_user_identity() const;
     void set_current_user_identity(const std::string& identity);
-
-    util::Optional<SyncAppMetadata> get_app_metadata();
-    /// Set or update the cached app server metadata. The metadata will not be updated if it has already been
-    /// set and the provided values are not different than the cached information. Returns true if the metadata
-    /// was updated.
-    /// @param deployment_model The deployment model reported by the app server
-    /// @param location The location name where the app server is located
-    /// @param hostname The hostname to use for the app server admin api
-    /// @param ws_hostname The hostname to use for the app server websocket connections
-    bool set_app_metadata(const std::string& deployment_model, const std::string& location,
-                          const std::string& hostname, const std::string& ws_hostname);
 
     /// Construct the metadata manager.
     ///
@@ -257,14 +236,12 @@ private:
     Realm::Config m_metadata_config;
     SyncUserMetadata::Schema m_user_schema;
     SyncFileActionMetadata::Schema m_file_action_schema;
-    SyncAppMetadata::Schema m_app_metadata_schema;
 
     std::shared_ptr<Realm> get_realm() const;
     std::shared_ptr<Realm> try_get_realm() const;
     std::shared_ptr<Realm> open_realm(bool should_encrypt, bool caller_supplied_key);
 
-
-    util::Optional<SyncAppMetadata> m_app_metadata;
+    bool run_file_action(SyncFileManager& file_manager, SyncFileActionMetadata& md) const;
 };
 
 } // namespace realm